Experimental implementation for event payloads
This implements Event payloads using traits. The method Event::data<T> returns a
reference to EventPayload<T>, and the struct EventPayload<T> defines the only
one or two method that are used inside of the container code (unfolding / event
/ cut / configuration).

In particular, Event::str() is now a template Event::str<T>(), because it needs
to access the payload to print it. This in turn makes that many other str()
methods in the container become templates. Since they become templates they need
to be in the .hpp file instead of the .cc. And moving them to the .hpp triggered
a significant reorganization of the code.